Set URL on first time comments loaded.
Previously, `reset` was only run if the disqus code was already loaded.  This was fine before the component took on the responsibility of setting the URL, but since the `reset` function is the only place that the URL is set it now must be run every time, even the first.

This does add another problem- `reset` being run before `window.DISQUS` is loaded.  This is solved by checking whether it is loaded, and re-running the function every 100 milliseconds until `window.DISQUS` is properly loaded.

If there is a more elegant way of setting the page url, please advise.